moving  columns to molecules directory
